# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "filex - files and agents API",
    "version": "0.50.0",
    "description": "The routes under `/api/files` (what the web explorer, the desktop app, embeds and `filex client` call) and `/api/ai` (the agent surface: REST and the MCP endpoint). Every operation here is a route the server registers: `web/tests/docs/openapiRoutes.test.ts` holds this file to `backend/internal/api/routes.go` in both directions - no operation the server does not serve, no `/api/files` or `/api/ai` route left out. The admin API (`/api/admin`, `/api/ai/admin`) is not described here; see docs/BACKEND.md.\n\nPaths of files are `<storage>://<path>`. A change sent by a browser with only the session cookie from another origin is refused (`403 cross_origin_refused`); a Bearer or `X-Filex-Token` header is not affected. An API token is held to its verbs (`x-filex-token-verb` on `/api/files`, `x-filex-token-scope` on `/api/ai`), and where an operation also asks a permission with a level, to that (`x-filex-token-permission`: `comments:read`, `comments:write` - the token's `comments:rw`; docs/RBAC.md).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "The filex server itself."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenHeader": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/ai/notifications/read": {
      "post": {
        "operationId": "aiNotificationRead",
        "summary": "Mark one notification, or all of them, read",
        "tags": [
          "Agents (/api/ai)"
        ],
        "description": "Runs the explorer's own handler for the token's user (handlers/ai_doors.go) and answers with that handler's status and body; the REST twin of an MCP tool, asking the tool's scope.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "all": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Marked read. The bell's own answer (POST /api/notifications/{id}/read, /read-all): no body, and the same 204 for an id the caller cannot see."
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tokenHeader": []
          }
        ],
        "externalDocs": {
          "url": "https://github.com/BRF-Tech/filex/blob/main/docs/MCP.md"
        },
        "x-filex-token-scope": "read"
      }
    },
...
    "/api/files/plugins/ui/{plugin}/{view}/read": {
      "get": {
        "operationId": "appUiRead",
        "summary": "An app interface reads the file it was opened with",
        "description": "Checked like a save: the app is running, the view is its interface, the person may run apps (`plugins.run`), the app was granted `files:read`, and the administrator did not turn the app off for this kind of file. The read itself is the explorer's preview read - the person's read permission, confinement and ranges - so an app reads nothing its person could not.",
        "tags": [
          "Apps"
        ],
        "parameters": [
          {
            "name": "plugin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The file, `<storage>://<path>`."
          }
        ],
        "responses": {
          "200": {
            "description": "The file's bytes (a `Range` header is honoured).",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-filex-token-verb": "read"
      }
    },
...
    "/api/files/read": {
      "get": {
        "operationId": "readFile",
        "summary": "A file's bytes by id (or storage + path)",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "The node id (preferred)."
          },
          {
            "name": "storage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "The storage id, with `path`."
          },
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "The path inside the storage."
          },
          {
            "name": "download",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "`1`: as an attachment."
          }
        ],
        "responses": {
          "200": {
            "description": "The bytes.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "externalDocs": {
          "url": "https://github.com/BRF-Tech/filex/blob/main/docs/BACKEND.md#get-apifilesread"
        },
        "x-filex-token-verb": "read"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A session token (`POST /api/auth/login`) or an API token (`fxt_...`)."
      },
      "tokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Filex-Token",
        "description": "An API token."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "filex_session",
        "description": "The web app's session."
      }
    },
    "responses": {
      "Error": {
        "description": "A refusal or a failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "The refusal's code: lower-case, stable across releases (`read_only`, `quota_exceeded`, `permission_denied`, ...). Some older routes still put an English sentence here; branch on the code, print `message` (docs/API-ERRORS.md)."
          },
          "code": {
            "type": "string",
            "description": "A second, upper-case code some routes have always sent beside `error` (`NAME_TAKEN`, `READ_ONLY`, `E2E_ENCRYPTED`, ...); kept for the clients that read it."
          },
          "message": {
            "type": "string",
            "description": "The sentence to show a person, written by the server in the reader's language (the account's, else Accept-Language). Print it as it is."
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "The values `message` was filled with (a name, a limit), for a program that wants them."
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}